fix(core): read opcodes as full 32-bit, not truncated ushort
TW opcodes carry a category byte in the upper 16 bits (e.g. 0x00021208
buff-state, 0x0001FBD4 pet). Narrowing to ushort collided these with
0x1208 / 0xFBD4. Aura's Op is int for the same reason.

Changed Op from ushort to uint across the framer, model, decoder
registry, filters, JSON envelope and all 24 decoders. Added a framer
regression test asserting 0x00021208 survives without truncation.

+++ src/Mabipacade.Core/Diagnostics/SessionEvent.cs
@@ -16,8 +16,8 @@ public abstract record SessionEvent(DateTime TimestampUtc)
         : SessionEvent(TimestampUtc);
     public sealed record FrameResync(DateTime TimestampUtc, long ByteOffset, string Reason)
         : SessionEvent(TimestampUtc);
-    public sealed record BadBody(DateTime TimestampUtc, ushort Op, int Length)
+    public sealed record BadBody(DateTime TimestampUtc, uint Op, int Length)
         : SessionEvent(TimestampUtc);
-    public sealed record DecoderFailed(DateTime TimestampUtc, ushort Op, string ExceptionMessage)
+    public sealed record DecoderFailed(DateTime TimestampUtc, uint Op, string ExceptionMessage)
         : SessionEvent(TimestampUtc);
 }
