+++ src/Mabipacade.Decoders/Entity/EntityDisappearDecoder.cs
@@ -6,6 +6,6 @@ public sealed record EntityDisappear;
 
 public sealed class EntityDisappearDecoder : IPacketDecoder
 {
-    public ushort Op => 0x520D;
+    public uint Op => 0x520D;
     public object Decode(DecoderInput input) => new EntityDisappear();
 }
