+++ src/Mabipacade.Decoders/Misc/EffectDecoder.cs
@@ -6,6 +6,6 @@ public sealed record Effect;
 
 public sealed class EffectDecoder : IPacketDecoder
 {
-    public ushort Op => 0x9091;
+    public uint Op => 0x9091;
     public object Decode(DecoderInput input) => new Effect();
 }
