+++ src/Mabipacade.Decoders/Misc/EffectDelayedDecoder.cs
@@ -6,6 +6,6 @@ public sealed record EffectDelayed;
 
 public sealed class EffectDelayedDecoder : IPacketDecoder
 {
-    public ushort Op => 0x9095;
+    public uint Op => 0x9095;
     public object Decode(DecoderInput input) => new EffectDelayed();
 }
