+++ src/Mabipacade.Decoders/Misc/PartyWindowUpdateDecoder.cs
@@ -6,6 +6,6 @@ public sealed record PartyWindowUpdate;
 
 public sealed class PartyWindowUpdateDecoder : IPacketDecoder
 {
-    public ushort Op => 0xA43C;
+    public uint Op => 0xA43C;
     public object Decode(DecoderInput input) => new PartyWindowUpdate();
 }
