+++ src/Mabipacade.Decoders/Skills/PlayerSkillPrepareReadyDecoder.cs
@@ -6,7 +6,7 @@ public sealed record PlayerSkillPrepareReady(ushort SkillId);
 
 public sealed class PlayerSkillPrepareReadyDecoder : IPacketDecoder
 {
-    public ushort Op => 0x6985;
+    public uint Op => 0x6985;
     public object Decode(DecoderInput input) =>
         new PlayerSkillPrepareReady(input.Elems[0].AsUInt16());
 }
