+++ src/Mabipacade.Decoders/Stats/StatUpdatePublicDecoder.cs
@@ -6,6 +6,6 @@ public sealed record StatUpdatePublic;
 
 public sealed class StatUpdatePublicDecoder : IPacketDecoder
 {
-    public ushort Op => 0x7532;
+    public uint Op => 0x7532;
     public object Decode(DecoderInput input) => new StatUpdatePublic();
 }
