+++ tests/Mabipacade.Core.Tests/Plugins/DecoderInputTests.cs
@@ -13,7 +13,7 @@ public class DecoderInputTests
         var input = new DecoderInput(ts, Direction.Inbound, 0x520C, 1UL, elems);
 
         Assert.Equal(ts, input.TimestampUtc);
-        Assert.Equal((ushort)0x520C, input.Op);
+        Assert.Equal((uint)0x520C, input.Op);
         Assert.Equal(1UL, input.EntityId);
         Assert.Same(elems, input.Elems);
     }
